requirements.txt
streamlit
transformers
torch
pandas
numpy
Pillow